Update name NewUploadFile.vue to UploadFile.vue and Move UploadFile.vue
